Move chunk_t into its own source files. Rename chunk_list.* files into chunk.*
