+++ src/align_asm_colon.cpp
@@ -10,7 +10,7 @@
 #include "align_asm_colon.h"
 
 #include "align_stack.h"
-#include "chunk_list.h"
+#include "chunk.h"
 
 
 void align_asm_colon(void)
