+++ src/align_func_proto.h
@@ -10,7 +10,7 @@
 #ifndef ALIGN_FUNC_PROTO_H_INCLUDED
 #define ALIGN_FUNC_PROTO_H_INCLUDED
 
-#include "chunk_list.h"
+#include "chunk.h"
 
 //! Aligns all function prototypes in the file.
 void align_func_proto(size_t span);
