+++ src/align_struct_initializers.cpp
@@ -10,7 +10,7 @@
 #include "align_struct_initializers.h"
 
 #include "align_init_brace.h"
-#include "chunk_list.h"
+#include "chunk.h"
 
 
 void align_struct_initializers(void)
