+++ src/align_tools.h
@@ -10,7 +10,7 @@
 #ifndef ALIGN_TOOLS_H_INCLUDED
 #define ALIGN_TOOLS_H_INCLUDED
 
-#include "chunk_list.h"
+#include "chunk.h"
 #include "uncrustify_types.h"
