+++ src/align_var_def_brace.h
@@ -10,7 +10,7 @@
 #ifndef ALIGN_VAR_DEF_BRACE_H_INCLUDED
 #define ALIGN_VAR_DEF_BRACE_H_INCLUDED
 
-#include "chunk_list.h"
+#include "chunk.h"
 
 /**
  * Scan everything at the current level until the close brace and find the
