+++ src/flag_decltype.cpp
@@ -4,7 +4,7 @@
  * @license GPL v2+
  */
 
-#include "chunk_list.h"
+#include "chunk.h"
 
 
 bool flag_cpp_decltype(chunk_t *pc)
