+++ src/flag_parens.h
@@ -8,7 +8,7 @@
 #ifndef FLAG_PARENS_H_INCLUDED
 #define FLAG_PARENS_H_INCLUDED
 
-#include "chunk_list.h"
+#include "chunk.h"
 
 
 /**
