+++ src/newlines.h
@@ -9,7 +9,7 @@
 #ifndef NEWLINES_H_INCLUDED
 #define NEWLINES_H_INCLUDED
 
-#include "chunk_list.h"
+#include "chunk.h"
 
 /**
  * Double the newline, if allowed.
