+++ src/quick_align_again.cpp
@@ -10,7 +10,7 @@
 #include "quick_align_again.h"
 
 #include "align_stack.h"
-#include "chunk_list.h"
+#include "chunk.h"
 
 
 void quick_align_again(void)
