+++ src/remove_duplicate_include.h
@@ -9,7 +9,7 @@
 #ifndef REMOVE_DUPLICATE_INCLUDE_H_INCLUDED
 #define REMOVE_DUPLICATE_INCLUDE_H_INCLUDED
 
-#include "chunk_list.h"
+#include "chunk.h"
 
 /**
  * @brief Remove duplicate include
