+++ tests/expected/cpp/02103-output.cpp
@@ -7,7 +7,7 @@
 
 #include "uncrustify_types.h"
 #include "prototypes.h"
-#include "chunk_list.h"
+#include "chunk.h"
 #include <cstring>
 #include <cstdlib>
